// src/types/user.ts
export interface User {
  acc_id?: string
  branch_ids?: Array<string>
  cellphone?: string
  code?: string
  department?: string
  descr?: string
  email?: string
  landphone?: string
  name?: string
  position?: string
  region_pairs?: Array<any>
  role_id?: string
  role_name?: string
  role_type?: string
  time_stamp?: string
  update_time_stamp?: string
  user_account?: string
  user_id?: string
}